Move apps/constants to mkt/constants (bug 1021196)
