+++ apps/amo/tests/__init__.py
@@ -38,12 +38,12 @@ from waffle.models import Flag, Sample, Switch
 import amo
 import mkt
 from amo.urlresolvers import get_url_prefix, Prefixer, reverse, set_url_prefix
-from constants.applications import DEVICE_TYPES
 from lib.es.management.commands import reindex_mkt
 from lib.post_request_task import task as post_request_task
 from mkt.access.acl import check_ownership
 from mkt.access.models import Group, GroupUser
 from mkt.constants import regions
+from mkt.constants.applications import DEVICE_TYPES
 from mkt.files.helpers import copyfileobj
 from mkt.files.models import File, Platform
 from mkt.prices.models import AddonPremium, Price, PriceCurrency
