+++ mkt/constants/__init__.py
@@ -5,7 +5,7 @@ from .carriers import CARRIER_CHOICES, CARRIER_SLUGS
 from .categories import (CATEGORY_CHOICES, CATEGORY_CHOICES_DICT,
                          TARAKO_CATEGORY_CHOICES, TARAKO_CATEGORIES_MAPPING)
 from .features import APP_FEATURES, PRERELEASE_PERMISSIONS
-from .platforms import DEVICE_LOOKUP, FREE_PLATFORMS, PAID_PLATFORMS
+from .platforms import FREE_PLATFORMS, PAID_PLATFORMS, PLATFORMS_NAMES
 from .regions import REGIONS_CHOICES_ID
 from .submit import (APP_PREVIEW_MINIMUMS, APP_PREVIEW_SIZES,
                      MAX_PACKAGED_APP_SIZE)
