+++ mkt/constants/applications.py
