+++ mkt/developers/tests/test_forms_payments.py
@@ -6,7 +6,7 @@ from test_utils import RequestFactory
 
 import amo
 import amo.tests
-from constants.payments import (PAYMENT_METHOD_ALL, PAYMENT_METHOD_CARD,
+from mkt.constants.payments import (PAYMENT_METHOD_ALL, PAYMENT_METHOD_CARD,
                                 PAYMENT_METHOD_OPERATOR)
 from mkt.developers import forms_payments, models
 from mkt.developers.providers import get_provider
@@ -15,8 +15,8 @@ from mkt.developers.tests.test_views_payments import setup_payment_account
 from mkt.prices.models import AddonPremium, Price
 from mkt.reviewers.models import RereviewQueue
 from mkt.site.fixtures import fixture
-from mkt.webapps.models import Addon, AddonDeviceType, AddonUser
 from mkt.users.models import UserProfile
+from mkt.webapps.models import Addon, AddonDeviceType, AddonUser
 
 
 class TestPremiumForm(amo.tests.TestCase):
