+++ mkt/developers/tests/test_models.py
@@ -9,16 +9,15 @@ from nose.tools import eq_, ok_
 
 import amo
 import amo.tests
-from constants.payments import PROVIDER_BANGO, PROVIDER_BOKU
+from mkt.constants.payments import PROVIDER_BANGO, PROVIDER_BOKU
 from mkt.developers.models import (ActivityLog, ActivityLogAttachment,
                                    AddonPaymentAccount, CantCancel,
                                    PaymentAccount, PreloadTestPlan,
                                    SolitudeSeller)
 from mkt.developers.providers import get_provider
 from mkt.site.fixtures import fixture
-from mkt.webapps.models import Addon, Webapp
 from mkt.users.models import UserProfile
-
+from mkt.webapps.models import Addon, Webapp
 from .test_providers import Patcher
